# Optimize pool