Master:

|       Method | Pending |     Mean | Allocated | Survived |
|------------- |-------- |---------:|----------:|---------:|
|   AsyncAwait |    True | 2.147 us |         - |    696 B |
| ContinueWith |    True | 2.178 us |         - |    400 B |

This PR:

|       Method | Pending |     Mean | Allocated | Survived |
|------------- |-------- |---------:|----------:|---------:|
|   AsyncAwait |    True | 1.993 us |         - |    696 B |
| ContinueWith |    True | 1.942 us |         - |    400 B |